Moves non-gas Q&A to new section.
The first 3 questions in the Gas Fees section are not about gas fees, which can cause confusion to readers.  I moved them to a new FAQ section, but there may be a better way to handle this.

One question is about how Interface and Contract fees interact, so it doesn't seem quite right for either section.  One is a Q&A about Interface fees specifically so it could be moved into that section, but currently that section is not in Q&A form.

It is also worth noting that the Gas Fees section is in Q&A format while the other two sections are in a more prose-like format, it may be worth making the Gas Fees section format match the Interface and Contract fees sections in writing style.